refactor(cli): split tui.rs into tui/ module + extract TUI surface from main.rs
Final tier-1 god-file refactor. Two structural changes that together
shrink main.rs from 1674 lines to 527 lines (-69%) while establishing
a clean tui/ module that lives alongside the existing tui_export.rs,
tui_formatter.rs, and tui_settings.rs.

File-level changes:

  apps/netscli-cli/src/tui.rs                 →  tui/state.rs
  (the existing 2226-line ratatui app + TuiApp impl + render helpers,
  moved unchanged via 'git mv' so blame/history are preserved)

  apps/netscli-cli/src/tui/mod.rs             NEW  (415 lines)
  Public entry point. Carries 'pub async fn run_tui(...)' (extracted
  from main.rs) plus the 'TerminalCleanup' Drop guard that restores
  raw mode on TUI exit. Re-exports the public API surface
  (TuiApp, EntryState, HistoryEntry, help_lines) from state.rs so
  external call sites in main.rs and tui_export.rs keep working
  without touching their imports.

  apps/netscli-cli/src/tui/events.rs          NEW  (780 lines)
  'pub async fn handle_command' (renamed from handle_tui_command in
  main.rs). The slash-command dispatcher invoked from run_tui's event
  loop on '/discover', '/scan', '/inspect', '/sweep', '/dns', etc.

The two functions moved out of main.rs (run_tui at 366 lines and
handle_tui_command at 754 lines = 1120 total) lived there for
historical reasons but were entirely TUI-specific and had no callers
outside the TUI surface. Co-locating them with state.rs makes the TUI
module self-contained: external code only depends on
tui::run_tui(...) and the re-exported types.

Behavior is preserved exactly. handle_tui_command's body and signature
are unchanged — only the name (handle_command, since it's now in the
events module) and the call site in run_tui (events::handle_command).

main.rs now contains only:
- Module declarations + imports
- OutputFormat enum + output_format()/print_structured() helpers
- main() function (CLI arg dispatch, ~440 lines)
- Single 'tui::run_tui(...)' call when no subcommand is provided

Verified locally:
- cargo check -p netscli clean
- cargo clippy -p netscli --all-targets -- -D warnings clean
- cargo fmt --all -- --check clean
- cargo test -p netscli — all 25 tests pass
