+++ apps/netscli-cli/src/tui/state.rs
